backend/requirements.txt
fastapi>=0.110,<1
uvicorn[standard]>=0.23,<1
python-socketio[asgi]>=5.13.0
httpx>=0.24,<1
python-dotenv>=1,<2
requests>=2.31,<3
beautifulsoup4>=4.12,<5
lxml>=4.9,<5
aiofiles>=23.0,<24
python-multipart>=0.0.6,<1
openai>=1.40,<2